// src/thor/model/OasObjectSchema.ts
export type OasObjectSchema  = {
    /**
     * 
     * @type {string}
     * @memberof OasObjectSchema
     */
    oasComponentsId?: string;
    /**
     * 
     * @type {string}
     * @memberof OasObjectSchema
     */
    oasParameterId?: string;
    /**
     * 
     * @type {string}
     * @memberof OasObjectSchema
     */
    oasResponseId?: string;
    /**
     * 
     * @type {string}
     * @memberof OasObjectSchema
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof OasObjectSchema
     */
    format?: string;
    /**
     * 
     * @type {Array<OasEnum>}
     * @memberof OasObjectSchema
     */
    oasEnum?: Array<OasEnum>;
    /**
     * 
     * @type {Array<OasRequired>}
     * @memberof OasObjectSchema
     */
    required?: Array<OasRequired>;
    /**
     * Unique identifier for object in the system
     * @type {string}
     * @memberof OasObjectSchema
     */
    id?: string;
    /**
     * UUID of owner of the object in the system
     * @type {string}
     * @memberof OasObjectSchema
     */
    ownerId?: string;
    /**
     * Date of object creation
     * @type {Date}
     * @memberof OasObjectSchema
     */
    createdDate?: Date;
    /**
     * Data, including hash of the key(s) used to encrypt this record.
     * @type {string}
     * @memberof OasObjectSchema
     */
    keyHash?: string;
    /**
     * Last user to access object
     * @type {string}
     * @memberof OasObjectSchema
     */
    lastAccessedById?: string;
    /**
     * Timestamp of last access of object
     * @type {Date}
     * @memberof OasObjectSchema
     */
    lastAccessedDate?: Date;
    /**
     * Unique identifier for user who last modifed the object in the system
     * @type {string}
     * @memberof OasObjectSchema
     */
    lastModifiedById?: string;
    /**
     * Date of last object modification
     * @type {Date}
     * @memberof OasObjectSchema
     */
    lastModifiedDate?: Date;
}